+++ packages/dashboard/src/components/Wizard/index.ts
@@ -1,2 +1,5 @@
 export * from './SetupWizard';
 export * from './SetupWizard.types';
+
+export * from './PanelWizard';
+export * from './PanelWizard.types';
